remove all error code, insert example code
