+++ apps/backend/index.ts
@@ -1,18 +1,6 @@
 // apps/user-service/src/index.ts
 import http from 'http';
-import { createApp } from './app';
-import dotenv from 'dotenv';
-import path from 'path';
-
-dotenv.config({
-  path: path.resolve(__dirname, '../../../.env'), 
-});
-
-if (process.env.ENV_FILE) {
-  dotenv.config({
-    path: path.resolve(__dirname, '../../../', process.env.ENV_FILE),
-  });
-}
+import { createApp } from './app'
 
 const PORT = process.env.PORT || 3002;
 
@@ -21,13 +9,11 @@ async function startServer() {
   const server = http.createServer(app);
 
   server.listen(PORT, () => {
-    console.log('User Service listening on port', PORT);
+    console.log('Aarti app backend listening on port', PORT);
   });
-    
-  
 }
 
 startServer().catch((err) => {
-  console.error('Error starting User Service:', err);
+  console.error('Error starting Aarti app backend:', err);
 });
