+++ apps/backend/routes.ts
@@ -1,0 +1,12 @@
+// apps/backend/routes.ts
+import { Router } from 'express';
+import {
+    // import your controller logic here
+} from './controllers';
+
+const router = Router();
+
+// router.get('/', doSomethingController as any)
+// will return the result of your doSomethingController
+
+export default router;
