+++ apps/backend/services.ts
@@ -1,0 +1,12 @@
+export async function nothing() {} // to make it a module => if you add a function, you can remove
+
+/*
+export async function doSomething(you: User) {
+  try {
+    console.log('haha, gonna put you in the database')
+    database.enter(user)
+  } catch (error) {
+    console.log("error putting you in the database", error) // DB_STORAGE_EXCEEDED
+  }
+}
+*/
